+++ packages/components/src/integrations/react-hook-form/components/Form/index.ts
@@ -1,2 +1,6 @@
 export { default } from "@/integrations/react-hook-form/components/Form/Form";
 export { Form, type FormOnSubmitHandler, type FormProps } from "./Form";
+export {
+  type WithFormSubmitControllerProps,
+  useFormSubmitController,
+} from "./hooks/useFormSubmitController";