+++ packages/remote-elements/src/auto-generated/RemoteCodeBlockElement.ts
@@ -11,25 +11,11 @@ export class RemoteCodeBlockElement extends FlowRemoteElement<RemoteCodeBlockEle
 
   static override get remoteProperties() {
     return {
-      CodeTag: {},
-      PreTag: {},
       className: {},
       code: {},
-      codeTagProps: {},
-      color: {},
       copyable: {},
-      customStyle: {},
       language: {},
-      lineNumberContainerStyle: {},
-      lineNumberStyle: {},
-      lineProps: {},
-      renderer: {},
-      showInlineLineNumbers: {},
       showLineNumbers: {},
-      startingLineNumber: {},
-      useInlineStyles: {},
-      wrapLines: {},
-      wrapLongLines: {},
     };
   }